+++ token/services/identity/msp/idemix/id.go
@@ -12,7 +12,6 @@ import (
 
 	bccsp "github.com/IBM/idemix/bccsp/types"
 	"github.com/hyperledger-labs/fabric-smart-client/pkg/utils/proto"
-	"github.com/hyperledger-labs/fabric-smart-client/platform/view/driver"
 	m "github.com/hyperledger/fabric-protos-go/msp"
 	"github.com/hyperledger/fabric/msp"
 	"github.com/pkg/errors"
@@ -203,8 +202,4 @@ func (id *MSPSigningIdentity) Sign(msg []byte) ([]byte, error) {
 		return nil, err
 	}
 	return sig, nil
-}
-
-func (id *MSPSigningIdentity) GetPublicVersion() driver.Identity {
-	return id.MSPIdentity
 }
