+++ token/services/identity/msp/x509/ecdsa.go
@@ -9,19 +9,12 @@ package x509
 import (
 	"crypto/ecdsa"
 	"crypto/elliptic"
-	"crypto/rand"
-	"crypto/rsa"
 	"crypto/sha256"
 	"crypto/x509"
 	"encoding/asn1"
 	"encoding/pem"
 	"math/big"
 
-	"github.com/hyperledger-labs/fabric-smart-client/pkg/utils/proto"
-	"github.com/hyperledger-labs/fabric-smart-client/platform/view/driver"
-	"github.com/hyperledger-labs/fabric-smart-client/platform/view/view"
-	"github.com/hyperledger/fabric-protos-go/msp"
-	"github.com/hyperledger/fabric/bccsp/utils"
 	"github.com/pkg/errors"
 )
 
@@ -42,31 +35,15 @@ type ecdsaSignature struct {
 	R, S *big.Int
 }
 
-type edsaSigner struct {
-	sk *ecdsa.PrivateKey
-}
-
-func (d *edsaSigner) Sign(message []byte) ([]byte, error) {
-	dgst := sha256.Sum256(message)
-
-	r, s, err := ecdsa.Sign(rand.Reader, d.sk, dgst[:])
-	if err != nil {
-		return nil, err
-	}
-
-	s, _, err = ToLowS(&d.sk.PublicKey, s)
-	if err != nil {
-		return nil, err
-	}
-
-	return utils.MarshalECDSASignature(r, s)
-}
-
-type edsaVerifier struct {
+type ecdsaVerifier struct {
 	pk *ecdsa.PublicKey
 }
 
-func (d *edsaVerifier) Verify(message, sigma []byte) error {
+func NewECDSAVerifier(pk *ecdsa.PublicKey) *ecdsaVerifier {
+	return &ecdsaVerifier{pk: pk}
+}
+
+func (d *ecdsaVerifier) Verify(message, sigma []byte) error {
 	signature := &ecdsaSignature{}
 	_, err := asn1.Unmarshal(sigma, signature)
 	if err != nil {
@@ -99,73 +76,15 @@ func (d *edsaVerifier) Verify(message, sigma []byte) error {
 	return nil
 }
 
-func NewSigner() (view.Identity, driver.Signer, driver.Verifier, error) {
-	// Create ephemeral key and store it in the context
-	sk, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	pkRaw, err := PemEncodeKey(sk.Public())
-	if err != nil {
-		return nil, nil, nil, errors.Wrap(err, "failed marshalling public key")
-	}
-
-	mspSI := &msp.SerializedIdentity{
-		//Type:    msp.SerializedIdentity_PK,
-		IdBytes: pkRaw,
-	}
-	idRaw, err := proto.Marshal(mspSI)
-	if err != nil {
-		return nil, nil, nil, errors.Wrap(err, "failed marshalling msp serialized identity")
-	}
-
-	return idRaw, &edsaSigner{sk: sk}, &edsaVerifier{pk: &sk.PublicKey}, nil
-}
-
-func NewVerifier(pk *ecdsa.PublicKey) *edsaVerifier {
-	return &edsaVerifier{pk: pk}
-}
-
-func NewIdentityFromBytes(raw []byte) (view.Identity, driver.Verifier, error) {
-	mspSI := &msp.SerializedIdentity{}
-	err := proto.Unmarshal(raw, mspSI)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "failed unmarshalling to msp serialized identity")
-	}
-
-	genericPublicKey, err := PemDecodeKey(mspSI.IdBytes)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "failed parsing received public key")
-	}
-	publicKey, ok := genericPublicKey.(*ecdsa.PublicKey)
+// IsLowS checks that s is a low-S
+func IsLowS(k *ecdsa.PublicKey, s *big.Int) (bool, error) {
+	halfOrder, ok := curveHalfOrders[k.Curve]
 	if !ok {
-		return nil, nil, errors.New("expected *ecdsa.PublicKey")
+		return false, errors.Errorf("curve not recognized [%s]", k.Curve)
 	}
 
-	return raw, &edsaVerifier{pk: publicKey}, nil
-}
+	return s.Cmp(halfOrder) != 1, nil
 
-// PemEncodeKey takes a Go key and converts it to bytes
-func PemEncodeKey(key interface{}) ([]byte, error) {
-	var encoded []byte
-	var err error
-	var keyType string
-
-	switch key.(type) {
-	case *ecdsa.PrivateKey, *rsa.PrivateKey:
-		keyType = "PRIVATE"
-		encoded, err = x509.MarshalPKCS8PrivateKey(key)
-	case *ecdsa.PublicKey, *rsa.PublicKey:
-		keyType = "PUBLIC"
-		encoded, err = x509.MarshalPKIXPublicKey(key)
-	default:
-		err = errors.Errorf("Programming error, unexpected key type %T", key)
-	}
-	if err != nil {
-		return nil, err
-	}
-
-	return pem.EncodeToMemory(&pem.Block{Type: keyType + " KEY", Bytes: encoded}), nil
 }
 
 // PemDecodeKey takes bytes and returns a Go key
@@ -199,32 +118,4 @@ func PemDecodeKey(keyBytes []byte) (interface{}, error) {
 	}
 
 	return key, nil
-}
-
-// IsLow checks that s is a low-S
-func IsLowS(k *ecdsa.PublicKey, s *big.Int) (bool, error) {
-	halfOrder, ok := curveHalfOrders[k.Curve]
-	if !ok {
-		return false, errors.Errorf("curve not recognized [%s]", k.Curve)
-	}
-
-	return s.Cmp(halfOrder) != 1, nil
-
-}
-
-func ToLowS(k *ecdsa.PublicKey, s *big.Int) (*big.Int, bool, error) {
-	lowS, err := IsLowS(k, s)
-	if err != nil {
-		return nil, false, err
-	}
-
-	if !lowS {
-		// Set s to N - s that will be then in the lower part of signature space
-		// less or equal to half order
-		s.Sub(k.Params().N, s)
-
-		return s, true, nil
-	}
-
-	return s, false, nil
 }
